@@ -5,13 +5,13 @@ class LocationResult {
   /// The human readable name of the location. This is primarily the
   /// name of the road. But in cases where the place was selected from Nearby
   /// places list, we use the <b>name</b> provided on the list item.
-  String address; // or road
+  String? address; // or road
 
   /// Google Maps place ID
-  String placeId;
+  String? placeId;
 
   /// Latitude/Longitude of the selected location.
-  LatLng latLng;
+  LatLng? latLng;
 
   LocationResult({this.latLng, this.address, this.placeId});
 
