@@ -4,14 +4,14 @@ import 'package:google_maps_flutter/google_maps_flutter.dart';
 class NearbyPlace {
   /// The human-readable name of the location provided. This value is provided
   /// for [LocationResult.address] when the user selects this nearby place.
-  String name;
+  String? name;
 
   /// The icon identifying the kind of place provided. Eg. lodging, chapel,
   /// hospital, etc.
-  String icon;
+  String? icon;
 
   // Latitude/Longitude of the provided location.
-  LatLng latLng;
+  LatLng? latLng;
 
   @override
   String toString() {
