@@ -10,10 +10,10 @@ typedef WidgetBuilder<T> = Widget Function(BuildContext context, T snapshot);
 
 class FutureLoadingBuilder<T> extends StatefulWidget {
   const FutureLoadingBuilder({
-    Key key,
-    @required this.future,
+    Key? key,
+    required this.future,
     this.initialData,
-    @required this.builder,
+    required this.builder,
     this.mutable = false,
     this.loadingIndicator,
   })  : assert(builder != null),
@@ -35,22 +35,22 @@ class FutureLoadingBuilder<T> extends StatefulWidget {
   /// provided to the [builder] will become null, regardless of [initialData].
   /// (The error itself will be available in [AsyncSnapshot.error], and
   /// [AsyncSnapshot.hasError] will be true.)
-  final T initialData;
+  final T? initialData;
 
   /// default is true
   ///
   /// set to false if the future will change.
   final bool mutable;
 
-  final Widget loadingIndicator;
+  final Widget? loadingIndicator;
 
   @override
   _FutureLoadingBuilderState<T> createState() =>
       _FutureLoadingBuilderState<T>();
 }
 
-class _FutureLoadingBuilderState<T> extends State<FutureLoadingBuilder<T>> {
-  Future<T> future;
+class _FutureLoadingBuilderState<T> extends State<FutureLoadingBuilder<T?>> {
+  Future<T?>? future;
 
   @override
   void initState() {
@@ -60,7 +60,7 @@ class _FutureLoadingBuilderState<T> extends State<FutureLoadingBuilder<T>> {
 
   @override
   Widget build(BuildContext context) {
-    return FutureBuilder<T>(
+    return FutureBuilder<T?>(
       future: widget.mutable ? widget.future : future,
       initialData: widget.initialData,
       builder: (BuildContext context, AsyncSnapshot snapshot) {
