@@ -11,12 +11,12 @@ import 'package:flutter/material.dart';
 class S implements WidgetsLocalizations {
   const S();
 
-  static S current;
+  static S? current;
 
   static const GeneratedLocalizationsDelegate delegate =
       GeneratedLocalizationsDelegate();
 
-  static S of(BuildContext context) => Localizations.of<S>(context, S);
+  static S? of(BuildContext context) => Localizations.of<S>(context, S);
 
   @override
   TextDirection get textDirection => TextDirection.ltr;
@@ -33,7 +33,7 @@ class $en extends S {
   const $en();
 }
 
-class GeneratedLocalizationsDelegate extends LocalizationsDelegate<S> {
+class GeneratedLocalizationsDelegate extends LocalizationsDelegate<S?> {
   const GeneratedLocalizationsDelegate();
 
   List<Locale> get supportedLocales {
@@ -44,8 +44,8 @@ class GeneratedLocalizationsDelegate extends LocalizationsDelegate<S> {
   }
 
   LocaleListResolutionCallback listResolution(
-      {Locale fallback, bool withCountry = true}) {
-    return (List<Locale> locales, Iterable<Locale> supported) {
+      {Locale? fallback, bool withCountry = true}) {
+    return (List<Locale>? locales, Iterable<Locale> supported) {
       if (locales == null || locales.isEmpty) {
         return fallback ?? supported.first;
       } else {
@@ -55,29 +55,29 @@ class GeneratedLocalizationsDelegate extends LocalizationsDelegate<S> {
   }
 
   LocaleResolutionCallback resolution(
-      {Locale fallback, bool withCountry = true}) {
-    return (Locale locale, Iterable<Locale> supported) {
+      {Locale? fallback, bool withCountry = true}) {
+    return (Locale? locale, Iterable<Locale> supported) {
       return _resolve(locale, fallback, supported, withCountry);
     };
   }
 
   @override
-  Future<S> load(Locale locale) {
+  Future<S?> load(Locale locale) {
     final String lang = getLang(locale);
     if (lang != null) {
       switch (lang) {
         case "ar":
           S.current = const $ar();
-          return SynchronousFuture<S>(S.current);
+          return SynchronousFuture<S?>(S.current);
         case "en":
           S.current = const $en();
-          return SynchronousFuture<S>(S.current);
+          return SynchronousFuture<S?>(S.current);
         default:
         // NO-OP.
       }
     }
     S.current = const S();
-    return SynchronousFuture<S>(S.current);
+    return SynchronousFuture<S?>(S.current);
   }
 
   @override
@@ -89,7 +89,7 @@ class GeneratedLocalizationsDelegate extends LocalizationsDelegate<S> {
   ///
   /// Internal method to resolve a locale from a list of locales.
   ///
-  Locale _resolve(Locale locale, Locale fallback, Iterable<Locale> supported,
+  Locale _resolve(Locale? locale, Locale? fallback, Iterable<Locale> supported,
       bool withCountry) {
     if (locale == null || !_isSupported(locale, withCountry)) {
       return fallback ?? supported.first;
@@ -125,7 +125,7 @@ class GeneratedLocalizationsDelegate extends LocalizationsDelegate<S> {
         // If no country requirement is requested, check if this locale has no country.
         if (true != withCountry &&
             (supportedLocale.countryCode == null ||
-                supportedLocale.countryCode.isEmpty)) {
+                supportedLocale.countryCode!.isEmpty)) {
           return true;
         }
       }
@@ -136,6 +136,6 @@ class GeneratedLocalizationsDelegate extends LocalizationsDelegate<S> {
 
 String getLang(Locale l) => l == null
     ? null
-    : l.countryCode != null && l.countryCode.isEmpty
+    : l.countryCode != null && l.countryCode!.isEmpty
         ? l.languageCode
         : l.toString();
