@@ -15,7 +15,7 @@ import 'intl/messages_all.dart';
 class S {
   S();
   
-  static S current;
+  static late S current;
   
   static const AppLocalizationDelegate delegate =
     AppLocalizationDelegate();
@@ -31,7 +31,7 @@ class S {
     });
   } 
 
-  static S of(BuildContext context) {
+  static S? of(BuildContext context) {
     return Localizations.of<S>(context, S);
   }
 
