@@ -33,7 +33,7 @@ class RichSuggestion extends StatelessWidget {
     final List<TextSpan> result = [];
 
     String startText =
-        autoCompleteItem.text.substring(0, autoCompleteItem.offset);
+        autoCompleteItem.text!.substring(0, autoCompleteItem.offset);
     if (startText.isNotEmpty) {
       result.add(
         TextSpan(
@@ -47,8 +47,8 @@ class RichSuggestion extends StatelessWidget {
       );
     }
 
-    String boldText = autoCompleteItem.text.substring(autoCompleteItem.offset,
-        autoCompleteItem.offset + autoCompleteItem.length);
+    String boldText = autoCompleteItem.text!.substring(autoCompleteItem.offset!,
+        autoCompleteItem.offset! + autoCompleteItem.length!);
 
     result.add(
       TextSpan(
@@ -63,8 +63,8 @@ class RichSuggestion extends StatelessWidget {
 
     String remainingText = this
         .autoCompleteItem
-        .text
-        .substring(autoCompleteItem.offset + autoCompleteItem.length);
+        .text!
+        .substring(autoCompleteItem.offset! + autoCompleteItem.length!);
     result.add(
       TextSpan(
         text: remainingText,
